renamed HowToList and HowToListCard to HowToDoList and HowToDoListCard
